requirements.txt
Flask==2.3.3
Twilio==8.10.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-Talisman==1.0.0
Flask-CORS==4.0.0
Werkzeug==2.3.7
